server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
distro==1.9.0
django==5.1.5
django-cors-headers==4.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.2
openai==1.59.8
pdfminer-six==20231228
pdfplumber==0.11.5
pgvector==0.3.6
pillow==11.1.0
pip==24.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.5
pydantic-core==2.27.2
pymupdf==1.25.2
pypdfium2==4.30.1
python-dotenv==1.0.1
setuptools==75.8.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-extensions==4.12.2
